Refactored camera code
